core/mount: treat loop mounts like network filesystems
A loop mount could be mounted from a network filesystem
and there is no easy way to be certain.

If it is, the current dependencies create an ordering loop as the
loop mount looks like a "local" filesystem, which shouldn't be
unmounted until after local-fs.target, but it also must be
unmounted before a remote filesystem which much be unmounted
before the network is disabled, which is before local-fs.target.

To break this loop, we treat all loop mounts as "remote".
If important services require the content of loop mounts,
the might cause problems, but that seems unlikely.

Ideally the "remote/local" distinction would be richer, can could
be propagates from fs to dev to fs.