tmpfiles: be careful of non-btrfs filesystems
If btrfs_subvol_auto_qgroup() is called on a path that
leads to an automount point, it will trigger the automount,
which can hang if it is an NFS mount and the network isn't
up yet.

path_check_fstype() is more careful and will not trigger
the automount.

As an automount is definitely not wanted in this case,
check first with path_check_fstype().

Fixes #1959